# Faster array filter and array choose