--- a/src/fsharp/FSharp.Core/array.fs
+++ b/src/fsharp/FSharp.Core/array.fs
@@ -495,14 +495,14 @@ namespace Microsoft.FSharp.Collections
         [<CompiledName("Filter")>]
         let filter f (array: _[]) = 
             checkNonNull "array" array                        
-            let temp = Array.zeroCreateUnchecked array.Length : bool[]
+            let temp = Array.zeroCreateUnchecked array.Length
             let mutable c = 0
             for i = 0 to array.Length-1 do                
-            if f array.[i] then
-                temp.[i] <- true
-                c <- c + 1
+                if f array.[i] then
+                    temp.[i] <- true
+                    c <- c + 1
                     
-            let result = Array.zeroCreateUnchecked c : _[]
+            let result = Array.zeroCreateUnchecked c
             c <- 0    
             let mutable i = 0
             while c < result.Length do