## Summary
An Array.filter that is ~1.6 times faster and causes fer GC collections on average. However some cases will be up to 20% slower and cause more slightly more GC collections

## Testing Summary
Tested with BenchmarkDotNet, on int32 arrays of lengths 10, 100, 10000, 1 million and 10 million.  Each array length was tested with filter rates of 95%, 50%, and 5% with simple predicates (i.e. x > 95) and each combination tested with both a sorted and random array.

## Considerations
The size of the type could influence the relative results. I chose a 32bit type for the initial bench marking because that will be common. Results with 64 bit types are similar.  In theory wider types should do better, narrower types worse, I think.

## Test Data
```ini

Host Process Environment Information:
BenchmarkDotNet=v0.9.8.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-4712HQ CPU 2.30GHz, ProcessorCount=8
Frequency=2240907 ticks, Resolution=446.2479 ns, Timer=TSC
CLR=MS.NET 4.0.30319.42000, Arch=64-bit RELEASE [RyuJIT]
GC=Concurrent Workstation
JitModules=clrjit-v4.6.1590.0

Type=SIMDBenchmark  Mode=Throughput  Platform=X64  
Jit=RyuJit  GarbageCollection=Concurrent Workstation  

```
                Method |   Length |              Median |            StdDev | Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
---------------------- |--------- |-------------------- |------------------ |------ |------ |------ |------------------- |
  **NewFilterHalf_Sorted** |       **10** |          **48.8852 ns** |         **1.9493 ns** |  **0.00** |     **-** |     **-** |              **43.77** |
     FilterHalf_Sorted |       10 |         115.9892 ns |        21.1080 ns |  0.00 |     - |     - |              92.84 |
  NewFilterHalf_Random |       10 |          49.0833 ns |         1.7518 ns |  0.00 |     - |     - |              45.39 |
     FilterHalf_Random |       10 |         112.2629 ns |        20.7312 ns |  0.00 |     - |     - |              81.52 |
 NewFilter95Pct_Sorted |       10 |          36.4150 ns |         5.3642 ns |  0.00 |     - |     - |              44.99 |
    Filter95Pct_Sorted |       10 |          52.0051 ns |         8.4022 ns |  0.00 |     - |     - |              52.34 |
 NewFilter95Pct_Random |       10 |          36.5451 ns |         4.0880 ns |  0.00 |     - |     - |              44.32 |
    Filter95Pct_Random |       10 |          52.1776 ns |         9.0984 ns |  0.00 |     - |     - |              51.93 |
  NewFilter5Pct_Sorted |       10 |          47.6038 ns |         0.5386 ns |  0.00 |     - |     - |              48.93 |
     Filter5Pct_Sorted |       10 |         163.6424 ns |         9.6942 ns |  0.01 |     - |     - |             123.35 |
  NewFilter5Pct_Random |       10 |          47.8569 ns |         1.0111 ns |  0.00 |     - |     - |              48.59 |
     Filter5Pct_Random |       10 |         162.8680 ns |        15.5700 ns |  0.01 |     - |     - |             122.44 |
  **NewFilterHalf_Sorted** |      **100** |         **364.5118 ns** |         **4.6385 ns** |  **0.01** |     **-** |     **-** |             **160.75** |
     FilterHalf_Sorted |      100 |         603.4328 ns |        16.3279 ns |  0.02 |     - |     - |             375.67 |
  NewFilterHalf_Random |      100 |         354.0028 ns |         5.3481 ns |  0.01 |     - |     - |             168.28 |
     FilterHalf_Random |      100 |         565.7060 ns |        26.1788 ns |  0.02 |     - |     - |             373.84 |
 NewFilter95Pct_Sorted |      100 |         312.1971 ns |         1.6338 ns |  0.00 |     - |     - |              91.96 |
    Filter95Pct_Sorted |      100 |         353.1542 ns |        30.9495 ns |  0.00 |     - |     - |              92.74 |
 NewFilter95Pct_Random |      100 |         337.0689 ns |        27.4325 ns |  0.00 |     - |     - |              90.64 |
    Filter95Pct_Random |      100 |         360.4190 ns |        32.3120 ns |  0.00 |     - |     - |              90.40 |
  NewFilter5Pct_Sorted |      100 |         373.5856 ns |         4.4249 ns |  0.01 |     - |     - |             238.50 |
     Filter5Pct_Sorted |      100 |         781.4070 ns |        46.8054 ns |  0.03 |     - |     - |             671.65 |
  NewFilter5Pct_Random |      100 |         390.6033 ns |        47.4583 ns |  0.01 |     - |     - |             262.57 |
     Filter5Pct_Random |      100 |         781.8424 ns |        80.4807 ns |  0.04 |     - |     - |             708.98 |
  **NewFilterHalf_Sorted** |    **10000** |      **30,556.8400 ns** |       **210.9701 ns** |  **0.67** |     **-** |     **-** |          **14,593.75** |
     FilterHalf_Sorted |    10000 |      43,509.5633 ns |       285.6323 ns |  2.08 |     - |     - |          49,214.32 |
  NewFilterHalf_Random |    10000 |     103,060.7084 ns |     2,557.9599 ns |  0.67 |     - |     - |          14,721.99 |
     FilterHalf_Random |    10000 |      88,873.8444 ns |       953.0232 ns |  1.90 |     - |     - |          45,016.63 |
 NewFilter95Pct_Sorted |    10000 |      27,415.7550 ns |       445.5048 ns |  0.28 |     - |     - |           7,356.43 |
    Filter95Pct_Sorted |    10000 |      28,079.7067 ns |       357.9169 ns |  0.17 |     - |     - |           3,777.83 |
 NewFilter95Pct_Random |    10000 |      36,382.7935 ns |       880.6256 ns |  0.27 |     - |     - |           7,142.76 |
    Filter95Pct_Random |    10000 |      32,211.2855 ns |       577.6430 ns |  0.15 |     - |     - |           3,346.98 |
  NewFilter5Pct_Sorted |    10000 |      33,194.3436 ns |     2,557.3665 ns |  1.01 |     - |     - |          19,970.78 |
     Filter5Pct_Sorted |    10000 |      58,363.5794 ns |     2,490.8887 ns |  3.75 |     - |     - |          83,177.14 |
  NewFilter5Pct_Random |    10000 |      41,400.9849 ns |       367.5454 ns |  1.03 |     - |     - |          20,286.77 |
     Filter5Pct_Random |    10000 |      63,866.5587 ns |       721.7905 ns |  3.87 |     - |     - |          85,802.02 |
  **NewFilterHalf_Sorted** |  **1000000** |   **3,752,191.6349 ns** |    **73,705.8287 ns** |     **-** |     **-** | **14.91** |       **1,472,055.96** |
     FilterHalf_Sorted |  1000000 |   5,664,764.7303 ns |    98,872.6639 ns |  2.04 |     - | 27.92 |       2,655,489.50 |
  NewFilterHalf_Random |  1000000 |  11,407,309.0717 ns |   141,989.8433 ns |     - |     - | 12.37 |       1,579,398.35 |
     FilterHalf_Random |  1000000 |  10,626,131.2383 ns |   111,056.2512 ns |  2.08 |     - | 33.06 |       2,929,089.25 |
 NewFilter95Pct_Sorted |  1000000 |   2,982,007.8851 ns |    24,600.6136 ns |     - |     - |  4.36 |         411,855.74 |
    Filter95Pct_Sorted |  1000000 |   2,848,117.8324 ns |    24,588.1818 ns |  1.73 |     - |  1.75 |         289,815.73 |
 NewFilter95Pct_Random |  1000000 |   4,009,920.7040 ns |    43,517.4572 ns |     - |     - |  5.42 |         718,710.35 |
    Filter95Pct_Random |  1000000 |   3,384,835.4986 ns |    38,807.5234 ns |  2.23 |     - |  2.29 |         493,473.52 |
  NewFilter5Pct_Sorted |  1000000 |   4,509,582.3644 ns |    52,218.4732 ns |     - |     - | 19.09 |       2,108,188.59 |
     Filter5Pct_Sorted |  1000000 |   9,071,487.2930 ns |   115,799.5791 ns |  2.15 |     - | 34.10 |       5,602,160.36 |
  NewFilter5Pct_Random |  1000000 |   5,489,445.0941 ns |    49,693.9393 ns |     - |     - | 18.65 |       2,063,641.56 |
     Filter5Pct_Random |  1000000 |   9,606,113.1720 ns |   119,037.7597 ns |  1.65 |     - | 26.31 |       4,402,385.65 |
  **NewFilterHalf_Sorted** | **10000000** |  **34,377,263.3131 ns** |   **540,364.6940 ns** |     **-** |     **-** |  **1.93** |      **24,656,145.01** |
     FilterHalf_Sorted | 10000000 |  65,365,106.6288 ns | 4,181,122.7822 ns |     - |     - | 16.10 |      54,325,440.10 |
  NewFilterHalf_Random | 10000000 | 110,493,373.8888 ns | 1,003,579.0419 ns |     - |     - |  2.63 |      33,989,114.28 |
     FilterHalf_Random | 10000000 | 114,917,363.8175 ns | 1,067,398.1184 ns |     - |     - | 20.00 |      58,524,939.62 |
 NewFilter95Pct_Sorted | 10000000 |  28,453,545.8188 ns |   357,348.9620 ns |     - |     - |  0.98 |      16,746,241.81 |
    Filter95Pct_Sorted | 10000000 |  28,092,921.7500 ns |   707,483.5145 ns |     - |     - |  1.25 |      14,962,478.18 |
 NewFilter95Pct_Random | 10000000 |  40,729,462.4456 ns |   684,911.1497 ns |     - |     - |  1.42 |      17,799,392.29 |
    Filter95Pct_Random | 10000000 |  33,072,936.5381 ns |   416,483.0060 ns |     - |     - |  1.42 |      14,957,642.45 |
  NewFilter5Pct_Sorted | 10000000 |  45,520,937.5044 ns |   678,329.9672 ns |     - |     - | 12.70 |      27,212,054.43 |
     Filter5Pct_Sorted | 10000000 | 101,916,656.9613 ns | 2,425,007.4248 ns |     - |     - | 73.00 |      96,063,048.14 |
  NewFilter5Pct_Random | 10000000 |  49,494,077.6213 ns |   762,068.5359 ns |     - |     - |  2.00 |      39,527,540.10 |
     Filter5Pct_Random | 10000000 | 107,543,675.3950 ns | 1,473,536.0093 ns |     - |     - | 73.00 |      96,063,324.81 |