@jackmott 2016-08-09
Let me sort out the build errors, and also benchmark against a pre-allocated ResizeArray.  The tip about number of allocations being more important than bytes is a good one. I assume that is a better indicator of GC overhead that will be induced.

I did experiment with BitArray, and while it reduced total memory used, it was much slower.