@jackmott 2016-08-09
The solution to pre allocate the ResizeArray: `List<_>(array.Length)` has an average runtime  nearly identical to my PR, with twice as many GC collections.  Simpler code of course and still a net performance benefit.