@asik 2016-08-09
You could also just do away with the ResizeArray and manage an array and count yourself, and do the Array.Copy at the end (basically implementing what ResizeArray does for you, but without allocating the actual object). The increased GC could be due in part to those extra objects being allocated.