@jackmott 2016-08-09
Using the approach @asik suggested, which is:

```F#
let filterArray f (array: _[]) = 
    checkNonNull array
    let tmp = Array.zeroCreate array.Length 
    let mutable count = 0
    for i = 0 to array.Length - 1 do 
        let x = array.[i] 
        if f x then 
            tmp.[count] <- x
            count <- count + 1

    Array.sub tmp 0 count
```

I get faster run times than the boolean approach in this PR, but worse GC data than the current core lib function.  Not sure which tradeoff is better. To summarize the data below, for the large arrays (1 million):

* Current core lib filter: 6.8ms, 15.9 Gen 0 Collections, 141.9 Gen 2 Collections
* Bool Array Filter: 5.5ms, 107 Gen 2 Collections
* Array Filter: 4.7ms, 214 Gen 2 Collections


                   Method |  Length |             Median |          StdDev | Gen 0 | Gen 1 |  Gen 2 | Bytes Allocated/Op |
------------------------- |-------- |------------------- |---------------- |------ |------ |------- |------------------- |
   **BoolFilter_5Pct_Sorted** |     **100** |        **402.7818 ns** |       **1.9268 ns** |  **0.11** |     **-** |      **-** |             **253.96** |
  BoolFilter_50Pct_Sorted |     100 |        393.4204 ns |       3.3703 ns |  0.07 |     - |      - |             159.19 |
  BoolFilter_95Pct_Sorted |     100 |        341.9024 ns |       6.2344 ns |  0.04 |     - |      - |              86.70 |
   BoolFilter_5Pct_Random |     100 |        412.4110 ns |       8.3747 ns |  0.11 |     - |      - |             241.56 |
  BoolFilter_50Pct_Random |     100 |        391.6298 ns |       9.8372 ns |  0.07 |     - |      - |             150.25 |
  BoolFilter_95Pct_Random |     100 |        355.4542 ns |      27.4391 ns |  0.04 |     - |      - |              88.13 |
  ArrayFilter_5Pct_Sorted |     100 |        312.4544 ns |       6.1075 ns |  0.16 |     - |      - |             360.44 |
 ArrayFilter_50Pct_Sorted |     100 |        355.6463 ns |       7.6802 ns |  0.11 |     - |      - |             257.99 |
 ArrayFilter_95Pct_Sorted |     100 |        306.4356 ns |       7.0157 ns |  0.09 |     - |      - |             197.86 |
  ArrayFilter_5Pct_Random |     100 |        326.8768 ns |      10.3230 ns |  0.17 |     - |      - |             390.20 |
 ArrayFilter_50Pct_Random |     100 |        364.2332 ns |      13.1722 ns |  0.14 |     - |      - |             303.95 |
 ArrayFilter_95Pct_Random |     100 |        315.8146 ns |      12.4783 ns |  0.10 |     - |      - |             214.66 |
       Filter_5Pct_Sorted |     100 |        759.8227 ns |      12.7525 ns |  0.28 |     - |      - |             636.30 |
      Filter_50Pct_Sorted |     100 |        596.4863 ns |      20.8309 ns |  0.18 |     - |      - |             400.09 |
      Filter_95Pct_Sorted |     100 |        359.3931 ns |      26.9707 ns |  0.04 |     - |      - |              93.54 |
       Filter_5Pct_Random |     100 |        772.7896 ns |      42.6297 ns |  0.27 |     - |      - |             601.20 |
      Filter_50Pct_Random |     100 |        571.8679 ns |      17.0173 ns |  0.17 |     - |      - |             389.51 |
      Filter_95Pct_Random |     100 |        349.3997 ns |      20.7281 ns |  0.04 |     - |      - |              84.22 |
   **BoolFilter_5Pct_Sorted** | **1000000** |  **4,787,814.4318 ns** |  **31,994.3347 ns** |     **-** |     **-** | **175.35** |       **2,254,630.84** |
  BoolFilter_50Pct_Sorted | 1000000 |  3,967,269.1682 ns |  28,105.0362 ns |     - |     - | 113.40 |       1,351,977.00 |
  BoolFilter_95Pct_Sorted | 1000000 |  3,226,135.1095 ns |  97,420.8513 ns |     - |     - |  41.52 |         663,501.35 |
   BoolFilter_5Pct_Random | 1000000 |  5,670,363.7466 ns |  48,328.4452 ns |     - |     - | 159.38 |       2,063,570.88 |
  BoolFilter_50Pct_Random | 1000000 | 11,303,333.3156 ns | 181,749.2972 ns |     - |     - | 112.01 |       1,604,815.87 |
  BoolFilter_95Pct_Random | 1000000 |  4,229,216.6743 ns |  76,169.8371 ns |     - |     - |  43.92 |         685,068.24 |
  ArrayFilter_5Pct_Sorted | 1000000 |  5,071,080.7331 ns |  75,959.0832 ns |     - |     - | 248.22 |       3,431,623.16 |
 ArrayFilter_50Pct_Sorted | 1000000 |  3,952,696.3859 ns |  67,900.5575 ns |     - |     - | 155.75 |       2,757,462.50 |
 ArrayFilter_95Pct_Sorted | 1000000 |  2,798,249.6317 ns |  13,028.6719 ns |     - |     - | 260.24 |       2,232,151.29 |
  ArrayFilter_5Pct_Random | 1000000 |  5,478,145.9584 ns |  76,983.6737 ns |     - |     - | 252.40 |       3,487,852.82 |
 ArrayFilter_50Pct_Random | 1000000 |  7,648,695.6508 ns | 120,255.4384 ns |     - |     - | 153.00 |       2,858,622.99 |
 ArrayFilter_95Pct_Random | 1000000 |  3,210,554.7831 ns |  17,502.0476 ns |     - |     - | 219.48 |       1,925,062.52 |
       Filter_5Pct_Sorted | 1000000 |  8,914,289.5042 ns | 235,755.8701 ns | 11.94 |     - | 189.69 |       3,774,287.90 |
      Filter_50Pct_Sorted | 1000000 |  5,708,856.1127 ns | 130,737.9476 ns | 18.29 |     - | 262.33 |       2,770,996.23 |
      Filter_95Pct_Sorted | 1000000 |  2,902,162.6344 ns |  25,124.9486 ns | 21.65 |     - |  23.79 |         387,816.94 |
       Filter_5Pct_Random | 1000000 |  9,593,318.4086 ns |  94,689.1429 ns | 11.94 |     - | 189.69 |       3,774,024.28 |
      Filter_50Pct_Random | 1000000 | 10,312,837.3356 ns | 127,669.9123 ns | 12.26 |     - | 166.22 |       2,131,092.39 |
      Filter_95Pct_Random | 1000000 |  3,396,103.2575 ns | 122,485.1784 ns | 19.47 |     - |  19.97 |         500,099.13 |