@KevinRansom 2016-08-09
@jackmott.  I think we use lists + Add() quite a lot in fsharp.core.  I suspect this technique may be generally applicable in several other places.  I like the performance performance win 6.8 down to 4.7 is a pretty enormous raw perf gain.  We also many ... many fewer gen 2 collections , and likely a pattern of use we can replicate through F# core.

I think between you and @asik you have found us some dignificant performance gains.

This is great work ....

Kevin