@jackmott 2016-08-09
@KevinRansom  to clarify, the fast option that runs in 4.7 has more Gen 2 collections, not less.