@asik 2016-08-09
How come you're having that many Gen2 collections? Something must be keeping objects around longer than needs be, probably in your benchmark code somewhere? I would expect to see no Gen2 collections at all, only Gen0. Would like to see the benchmark code if you can share.